AdminResource- auth login, maps CRUD, person sb
backend auth login for restful api, maps CRUD
changed AdminResource methods to use PersonSB instead

+++ NUStartApplication/NUStartApplication-ejb/src/java/session/PersonSessionBean.java
@@ -1,5 +1,6 @@
 package session;
 
+import entity.Administrator;
 import entity.Category;
 import entity.Contact;
 import entity.Facility;
@@ -10,6 +11,7 @@ import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import enumeration.AccountType;
+import error.InvalidLoginException;
 import java.util.List;
 import javax.persistence.Query;
 
@@ -41,6 +43,25 @@ public class PersonSessionBean implements PersonSessionBeanLocal {
         return (Person) q.getSingleResult();
     }
 
+
+    //might be front end idk
+    @Override 
+    public Person administratorLogin(String username, String password) throws InvalidLoginException, NoResultException {
+        try {
+            //username is email
+            Person admin = getPersonByEmail(username);
+            if (admin.getPassword().equals(password)) {
+                return admin;
+            } else {
+                throw new InvalidLoginException("Administrator username does not exist or invalid password!");
+            }
+        } catch (InvalidLoginException ex) {
+            throw new InvalidLoginException("Administrator username does not exist or invalid password!");
+        } catch (NoResultException e) {
+            throw new NoResultException("Admin not found");
+        }
+    }
+    
     @Override
     public void createStaff(Person p) {
         p.setAccountType(AccountType.STAFF);
@@ -209,9 +230,17 @@ public class PersonSessionBean implements PersonSessionBeanLocal {
     }
 
     @Override
-    public void deleteMap(Long mId) {
+    public void deleteMap(Long mId) throws NoResultException {
         Map map = em.find(Map.class, mId);
-        em.remove(map);
+        try {
+            if (map != null) {
+                em.remove(map);
+            } else {
+                throw new NoResultException("Map not found");
+            }
+        } catch (NoResultException ex) {
+            throw new NoResultException("Map not found");
+        }
     }
 
     @Override
